Check

+++ 9-binary_tree_height.c
@@ -10,8 +10,9 @@ size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (!tree || (!tree->left && !tree->right))
 		return (0);
-	size_t left = binary_tree_height(tree->left);
-	size_t right = binary_tree_height(tree->right);
+	int left = binary_tree_height(tree->left);
+	int right = binary_tree_height(tree->right);
+
 	if (right > left)
 		return (1 + right);
 	else
